Add files via upload
This is the Google Data Analytics Certificate case study. Cyclistic is bike share company in Chicago. Lily Moreno, director of marketing, believes growth of number of annual members will be key to future growth. Moreno want to design the strategies to convert casual rider to annual members. The purpose of this analysis is to gain insight how annual member and casual riders use Cyclistic bike differently.